jordan: add page redirect and profile info
Implemented page redirection and profile information via react-cookies.

Page redirection works as follows.
  1) If not signed in, going to /profile redirects to /signin.
  2) If signed in, going to /signin or /signup redirects to /profile.
This is implemented with <Navigate to="path">, which is part of
react-router-dom.

Profile information and checking whether a user is logged in is done via
react-cookies, saved in the cookie name of "username".

The current implementation of cookies is wildly insecure and does not
have features like expiration dates, encryption, and whatnot. However,
that is future Jordan's problem :)