jordan2: generalize front/backend & db API
Revamped and generalized the communication 'api' between the frontend,
backend, and MongoDB.

Previously, separate route files in the backend and method files in the
frontend were created for each new collection in the database, which is
inefficient as they largely have the same actions. This has now been
generalized, where the new
  dbRoutes.js: link backend with MongoDB
  db.js: link frontend with backend (methods)
now handles any input object/objectID and collection name, with methods
  getObject: get object by ID (in database)
  createObject: add new object to database with given new object
  updateObject: update existing object with given updated object
  deleteObject: delete object by ID (in database)
(more can be added).

Similarly, the search.js (and searchRoutes.js) have been updated.

This is the generate function 'tempalte' for the new APIs.
  [action]Object(s) (id/object/search, collection, callback)
Where
  id/object/search: id, object, or search term depending on the function
  collection: name of the collection
  callback: (optional) function to be executed after the action has been
            done, typically used for page refreshes, default is
            basically 'do nothing'
Here are some examples.
  getObject("jordanlin", "users");
  createObject(newTile, "tiles", refresh);
  searchObjects("kay", "users");

Search.js, Signin.js, SignUp.js, and other files have been accordingly
updated to use the new APIs. Note that Profile.js is in a different
commit because it has other new stuff as well, so if reverting to this
exact commit Profile.js will be broken.