jordan2: add dedicated admin page
Added dedicated admin page (accessible only via manually going to /admin
in the URL bar) where admins (us!) can do admin stuff. The only current
feature is for us to upload new tiles, but more features can be easily
added, such as uploading other stuff.

Note that at this moment, to delete tiles, one must go to the MongoDB
interface and manually do it, with no image previews (so remember your
asset IDs/tags and whatnot).

Forgot to mention in the last commit, image uploads work with the
FileReader.readAsDataURL() function, where essentially images as read as
a stream of text that is displayable directly via
  <img src="yourImageStreamThingy" />
which is pretty convenient.