jordan: simply profile and user page to one
Simplified/merged profile and user pages to one master page. (Note that
there are some path and routing changes.) Now, one can go to
/user/[own-username] to view their own profile and make changes, while
/user/[some-other-username] to view others' profile without the ability
to make changes (of course). Hence, UserPage is deleted and the search
page is updated accordingly.

Moreover, the admin page is made visible to the five members of Eggmacs
in the navigation bar.

Also, I realized that react cookies can be accessed from any component
and do not need to be passed from the declaring parent component to
child components. (Of course this would be the case. That is the whole
point of cookies. Why did I not realized this.) Hence, now cookie
functions are no longer passed to child components in App.js but instead
child components that need cookies will just get them with useCookie.