jordan: establish data structure for map creator
Co-authored-by: Kay Rischpater <katie.risc@gmail.com>

Established and defined the data structure for the map creator, that is,
two overlayed 2D arrays. The asset array is 3 times the size (9 times
the number of entries) compared to the tile array as each tile square
contains 3 by 3 (9) asset squares.

Created the communication point (API) for creating new maps in the
profile page and updated MongoDB to match. Cookies and extra search
operations are used to display a list of maps in the profile page and
display the correct selected map information.

Map catalog display when viewing other users have yet to be supported.