Generalized Search to include maps, removed assetSearch
Because assets don't have a displayName in our database, their search
couldn't be generalized.  After some consideration, we decided it
was better to narrow search capabilities to maps and users.  Users
can already scroll through tiles in the mapCreator page (or rather,
will be able to); therefore, I made the search filer generalized
so that it should* work with the map collection :) going to
add that next! (also I removed the assetSearches from other portions of the code