Tile selection and display, Map creator functional!!
Implemented tile selection system

All tiles, assets, and characters are available to select in a bar underneath the map creator.

clicking on an icon in the asset switches tile placement to the selected asset

The icon of the tile being placed is outlined in red.

right click will now set grid square to empty

Map creation is now fully functional! yay win