Added Map Download Feature
After a painful ammount of wrestling with HTML5, I've added the
map download feature!  There is currently no way to call the
method, as my build was old enough that the button would conflict
with current CSS designs -- however, one simply needs to call
the handleMapDownload() and everything should work!

In order to download the map, there were two things that needed to
happen; first, we had to fetch all of the individual tiles, and
second, stitch them together into a single PNG.  Of course, there
were more complications than this; the long and short of my solution
is:

-Pass in the majority of Map info as an object, so that the
download.js method can process it

- Initialize an HTML5 Canvas to the size of the final map

- Iterate through the map, gathering the .body 64b datastreams for
each individual tile, and then each asset

- Convert the images to a JS Image(), and draw the tile onto the
canvas

- After the canvas object has been fully drawn, the b64 stream
is interpreted and converted to a .png file, and passed to
Creator.js to be downloaded using the FileSave.saveAs() feature!

The donwload.js has an additional "object to file" module; while
this is currently not being used, it can help download any images
from MongoDB we want in the future!