home page nav buttons + asset dump
For record purposes, all 51 assets I have currently drawn are uploaded and functional.

On the home page, finished final images for image buttons and uploaded them

Size of background images is scaled to match, cleaned up css for image buttons in general. The text over the buttons is now centered, unhyperlinked, and above the overlay layer.

Small cosmetic updates to the description as well, just so we don't have a hanging word at the end!